Seperating Security Config for Patient and Admin
