+++ admin-system/src/main/java/com/capstone/dentalclinic/demo/security/config/AdminSecurityConfig.java
@@ -1,6 +1,5 @@
 package com.capstone.dentalclinic.demo.security.config;
 
-import com.capstone.dentalclinic.demo.model.patient.Patient;
 import com.capstone.dentalclinic.demo.services.patient.PatientServicesImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -21,7 +20,7 @@ import lombok.AllArgsConstructor;
 @Configuration
 @EnableWebSecurity
 @AllArgsConstructor
-public class ApplicationSecurityConfig {
+public class AdminSecurityConfig {
     
     private final EmployeeServiceImpl employeeServiceImpl;
     private final PatientServicesImpl patientServicesImpl;
@@ -35,38 +34,8 @@ public class ApplicationSecurityConfig {
         return provider;
     }
 
-    @Bean
-    public DaoAuthenticationProvider daoAuthenticationProviderPatient() {
-        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
-        provider.setPasswordEncoder(bcryptPasswordEncoder);
-        provider.setUserDetailsService(patientServicesImpl);
-        return provider;
-    }
 
     @Bean
-    @Order(1)
-    public SecurityFilterChain securityFilterChainPatient (HttpSecurity http) throws  Exception{
-        http
-                .csrf().disable()
-                .authenticationProvider(daoAuthenticationProviderPatient())
-                .authorizeHttpRequests((authz) -> authz
-                        .antMatchers("/").permitAll()
-                        .antMatchers("/patient/*").permitAll()
-                        .antMatchers("/token/*").permitAll()
-                        .antMatchers("/patient/dashboard").hasRole("PATIENT")
-                        .anyRequest().authenticated()
-                )
-                .formLogin()
-                .loginPage("/patient/login")
-//                .usernameParameter()
-                .defaultSuccessUrl("/patient/dashboard", true)
-                .failureUrl("/admin/login-error");
-        return http.build();
-    }
-
-
-    @Bean
-    @Order(2)
     public SecurityFilterChain securityFilterChainAdministrator (HttpSecurity http) throws  Exception{
         http
             .csrf().disable()
